LayoutTests/svg/dom/SVGSVGElement-currentScale-NaN-no-crash.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="200" height="200"></svg>
<script>
description("Setting .currentScale to a non-finite should throw.");

self.jsTestIsAsync = true;

if (window.testRunner) {
     testRunner.dumpAsText();
     testRunner.waitUntilDone();
}

function runTest()
{
    svgElement = document.getElementById("svg");
    shouldThrow("svgElement.currentScale = NaN");
    shouldThrow("svgElement.currentScale = Infinity");
    finishJSTest();
}
window.onload = runTest;
</script>
</body>
</html>
